% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q7.R
\name{downloadRedcapReport}
\alias{downloadRedcapReport}
\title{Download a report from RedCap}
\usage{
downloadRedcapReport(redcapTokenName, redcapUrl, redcapReportId)
}
\arguments{
\item{redcapTokenName}{The name of the environment variable storing the RedCap API token.}

\item{redcapUrl}{The URL of the RedCap API endpoint.}

\item{redcapReportId}{The ID of the RedCap report to be retrieved.}
}
\value{
A tibble containing the data from the RedCap report.
}
\description{
Retrieves a RedCap report using the RedCap API, with parameters specified by the user.
The API token is fetched from the user's environment using Sys.getenv().
}
\examples{
# Make sure to set your environment variable before using this example
# Sys.setenv(REDCAP_TOKEN = "your_actual_token_here")
downloadRedcapReport("REDCAP_TOKEN", "https://redcap.yourinstitution.edu/api/", "12345")
}
