% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q6.R
\name{minimumN}
\alias{minimumN}
\title{Calculate Minimum Sample Size for a Two-Sample t-Test}
\usage{
minimumN(x1, x2 = NULL)
}
\arguments{
\item{x1}{A numeric vector containing the first sample of preliminary data.}

\item{x2}{An optional numeric vector containing the second sample of preliminary data. If not provided, a one-sample t-test scenario is assumed.}
}
\value{
A list containing the computed minimum sample sizes for each group.
}
\description{
This function calculates the minimum sample size needed to achieve 80\% power at a significance level of 0.05 for a two-sample t-test.
It uses the \code{pwr::pwr.t2n.test} function and can handle one or two samples of preliminary data.
}
\examples{
sample1 <- rnorm(20, mean = 10, sd = 2)
sample2 <- rnorm(20, mean = 12, sd = 2)
minimumN(sample1, sample2)
}
