% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q4.R
\name{pcApprox}
\alias{pcApprox}
\title{Principal Component Approximation of Data}
\usage{
pcApprox(x, npc)
}
\arguments{
\item{x}{A numeric matrix or data frame where rows are observations and columns are variables.}

\item{npc}{The number of principal components to use for the approximation.}
}
\value{
A numeric matrix of the same dimensions as \code{x}, approximated using the first \code{npc} principal components.
}
\description{
This function returns an approximation of the data matrix \code{x} based on the number of principal
components specified by \code{npc}. The approximation will be rescaled and recentered to match the
original data's scale and center.
}
\examples{
data(mtcars)
approx_data <- pcApprox(mtcars, 2)
print(approx_data)
}
