% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q5.R
\name{standardizeNames}
\alias{standardizeNames}
\title{Standardize Variable Names in a Tibble}
\usage{
standardizeNames(data)
}
\arguments{
\item{data}{A tibble or dataframe whose variable names you want to standardize.}
}
\value{
A tibble or dataframe with standardized variable names.
}
\description{
This function standardizes the variable names in a tibble or dataframe by first cleaning the names using
janitor::make_clean_names and then converting them to "small camel" case using snakecase::to_small_camel_case.
}
\examples{
library(tibble)
data <- tibble(`First name` = c("John", "Jane"), `Last name` = c("Doe", "Doe"), `AGE (years)` = c(28, 22))
standardized_data <- standardizeNames(data)
print(standardized_data)
}
