% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q2.R
\name{survCurv}
\alias{survCurv}
\title{Calculate and Plot a Survival Curve}
\usage{
survCurv(status, time)
}
\arguments{
\item{status}{Numerical vector of event occurrences (1 = event, 0 for censored).}

\item{time}{Numerical vector of times to event or censoring.}
}
\value{
Plot of the survival curve.
}
\description{
This function takes a numerical vector \code{status} indicating event occurrence
(1 for event, 0 for censored) and a numerical vector \code{time} indicating the
time to event or censoring, then calculates and plots the survival curve using
a non-parametric estimator.
}
\examples{
survCurv(c(1, 0, 1, 0, 1), c(5, 8, 12, 15, 20))
}
