% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q3.R
\name{unscale}
\alias{unscale}
\title{Unscale Function}
\usage{
unscale(x, center = NULL, scale = NULL)
}
\arguments{
\item{x}{Numeric vector to unscale.}

\item{center}{Numeric scalar indicating the centering factor (optional).}

\item{scale}{Numeric scalar indicating the scaling factor (optional).}
}
\value{
Numeric vector with centering and scaling reversed.
}
\description{
This function reverses the centering and scaling applied to a vector,
either by using provided center and scale values or by retrieving
the original values if the vector was scaled with R's scale() function.
}
\examples{
unscale(c(1, 2, 3), center = 2, scale = 1.5)
}
